+++ backend/models/inventory_item_in_composition.py
@@ -7,7 +7,7 @@ class InventoryItemInComposition(Base):
     id = Column(Integer, primary_key=True, index=True)
     composition_id = Column(Integer, ForeignKey("composition.id"))
     inventory_item_id = Column(Integer, ForeignKey("inventory_items.id"))
-    weight = Column(Float, nullable=False)
+    weight = Column(Numeric(10, 3), nullable=False)
     wastage_percentage = Column(Numeric(5, 2), nullable=True)
     tenant_id = Column(String, index=True)
     inventory_item = relationship("InventoryItem")
