+++ backend/schemas/composition_usage_history.py
@@ -1,12 +1,13 @@
 from pydantic import BaseModel
 from typing import Optional, List
 from datetime import datetime
+from decimal import Decimal
 
 # Schema for an item within a usage history record
 class CompositionUsageItem(BaseModel):
     item_name: str
     item_category: str
-    weight: float
+    weight: Decimal
 
     class Config:
         from_attributes = True
@@ -17,7 +18,7 @@ class CompositionUsageHistory(BaseModel):
     composition_id: int
     composition_name: str
     batch_id: int
-    times: float
+    times: int
     used_at: datetime
     items: List[CompositionUsageItem]
 
@@ -27,17 +28,17 @@ class CompositionUsageHistory(BaseModel):
 # Schema for the feed breakdown in the by-date response
 class FeedBreakdown(BaseModel):
     feed_type: str
-    amount: float
+    amount: Decimal
 
 # Schema for the by-date response
 class CompositionUsageByDate(BaseModel):
-    total_feed: float
+    total_feed: Decimal
     feed_breakdown: List[FeedBreakdown]
 
 class CompositionUsageCreate(BaseModel):
     compositionId: int
     batch_no: str
-    times: float
+    times: int
     usedAt: Optional[datetime] = None
